Create GMR_GraphingScript.py
This script is designed to read a series of excel files and plot reasonably modern
charts from the data. It is specifically for Ground Magnetic Resonance inversion data
so that the data can be visualised.

The script makes use of Pandas and ggplot in addition to matplotlib

Features:
    + the y axis has zero at the top because we are looking at depths below ground level
    + the depth of investigation horizontal line is determined by the data
    + pandas is used to filter the data were needed.
    + axis are named
    + all graphs are on the same x scale (40% max water content) to be comparable
    + dpi of the output png files is ajustable

Further improvement could be a smoothing function to take the jaggies out of the lines.